Crate pallet_transaction_payment_rpc_runtime_api

Runtime API definition for transaction payment pallet.

Structs

FeeDetails

The FeeDetails is composed of:

InclusionFee

The base fee and adjusted weight and length fees constitute the inclusion fee.

RuntimeDispatchInfo

Information related to a dispatchable’s class, weight, and fee that can be queried from the runtime.

Traits

TransactionPaymentApi

TransactionPaymentCallApi